+++ devops/push_nexus.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-echo "Push to Sonatype Nexus"
-
-wheel=$(find dist -name "*.whl")
-twine upload $wheel